fix bug, no upload new readings after internet failure
The bug that new readings from when out of internet are not uploaded to NS, even after internet

- as mentioned in the comments of func uploadData(dataToUpload: Any, httpMethod: String?, path: String, completionHandler: (() -> ())?)
completionhandler should be called only when upload was successful.
But the result of the upload was never checked. So probably, in some error cases, completionhandler was called with success. As a result, the upload was considered as successful and not redone when internet comes back

- als some updates in func uploadDataAndGetResponse, in some error cases, the completionhandler was not called. Although this has probably nothing to do with this bug, it may have caused 'hanging' treatment uploads